Deleted main project files
Doing a rework to better utilize a better project structure, and easier command structure.

New version should be able to work on multiple servers, and have less weird dependency system for each command. This will allow one command to work on all servers that this bot is added on. This will also help with testing. We can allow a test server to use for new features. Overall should be a better experience.